% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateASV.r
\name{aggregateASV}
\alias{aggregateASV}
\title{Aggregate ASVs to putative species}
\usage{
aggregateASV(x, by, other_str, other_num)
}
\arguments{
\item{x}{the columns containing number of reads}

\item{by}{The grouping variable}

\item{other}{Columns containing classification or other valuable information}
}
\description{
Several ASVs will be assigned to same species name. This function aggregates ASVs according to a variable (species name, accession number etc) 
Caution be sure than only aggregate relieable species assignments.
}
\details{

}
\examples{
data(deepMeio)
# grade >= 97 will give an indication of good species assignment
grade <- ((deepMeio$pident) + 2*(deepMeio$qcovs))/3
shortMeio <- deepMeio[grade>= 97,]
# grade > 97 will give an indication of good species assignment
# retrieve classification from WORMS
tax_deepMeio <- taxFromWorms(shortMeio$Species)

# aggregate ASVs  
agg_tax_deepMeio <- aggregateASV(x=shortMeio[,9:ncol(shortMeio)], by=shortMeio$Species,other_str=tax_deepMeio[,2:6],other_num=shortMeio[,2:5])

agg_tax_deepMeio_2 <- aggregateASV(x=shortMeio[,9:ncol(shortMeio)], by=shortMeio$Species,other_num=shortMeio[,2:5])

}
\seealso{
\code{\link{taxFromWorms}}
}
\author{
Pedro Martinez Arbizu
}
