% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASV_outliers.r
\name{ASV_outliers}
\alias{ASV_outliers}
\title{remove ASV outliers using observed distribution}
\usage{
ASV_outliers(x, by, n_var = 1)
}
\arguments{
\item{x}{The community table with Taxon reads counts}

\item{by}{A list with variables used to produce the threshold}

\item{n_var}{A number, minimum number of variables that should be below threshold to declare outliers default = 1.}
}
\value{
Two data.frames containing outliers and non outliers.
\item{keep}{data.frame with non outlier ASVs (above the treshold)} 
\item{outliers}{data.frame with outlier ASVs  (below the treshold)} 
\item{thresholds}{Table with thresholds used} 
\item{AVS_tm}{Table with thresholds matrix per ASV}
}
\description{
Recognize and remove ASV outliers according to selected variables
}
\details{
This function will detect outlier ASVs and divide the data.frame into outliers and not outliers accordingly. The threshold will be assigned by the lower wisker of the boxplot (1.5x interquartile variation).
Variables used to produced the treshhold a free configurable, but query coverage, percent identity and query length are meaningful.
User can define how many of the selected variables should be below the hreshold for the ASV to be declared as outlier. Default is 1, which correspond to combine the variables using an 'or' logical operator.
}
\examples{
ASV_outliers(x=deepMeio[,9:27],by=data.frame(deepMeio$qcovs,deepMeio$length,deepMeio$pident),n_var=1)




}
\seealso{
\code{\link{boxplot}}
}
\author{
Pedro Martinez Arbizu
}
