% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barp.table.r
\name{barp.table}
\alias{barp.table}
\title{Barplot Table}
\usage{
barp.table(x, by, fun, ...)
}
\arguments{
\item{x}{The community table with Taxon reads counts}

\item{by}{Aggregating vector of classes}

\item{fun}{Function used to aggregate. Default is sum.}
}
\value{
An object of class barp_table
A matrix with Taxa and aggregate values.
The object has plot method, which will produce barplots
For absolute and relative number of reads
and absolute and relative number of OTUs/ASV per Taxon
}
\description{
Creates a table that be be used for a barplot graph
}
\details{
This is a wrapper function to 'aggregate' and 'apply'.
It takes a Taxon table (OTU or ASV table) and counts the number of reads 
aggregating by the vector specified by 'by' using function specified with 'fun'
}
\examples{



}
\seealso{
\code{\link{plot.barp_table}}
}
\author{
Pedro Martinez Arbizu
}
