% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.dist.r
\name{check.dist}
\alias{check.dist}
\title{Check for Distribution}
\usage{
check.dist(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
Ranked solutions
}
\description{
Use MASS fitdistr function to find best fitting distribution
}
\details{
If x contains integer values, fitdistr is used to check for "negative binomial" or "Poisson"
distribution. If x is float following distributions are tested "lognormal",
"normal", "exponential","geometric",'gamma'. Solution are ranked according to 
AIC information criterium. Lowest is best.
}
\examples{
check.dist(rnorm(30,20,1)) 

# negative binomial distribution from a collection of poisson samples 
d <- c()
for(i in 1:20){d=c(d,rpois(50,sample(10:1000,1)))}
check.dist(d)




}
\seealso{
\code{\link{fitdistr}}
}
\author{
Pedro Martinez Arbizu
}
