% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonASV.r
\name{commonASV}
\alias{commonASV}
\title{Most common ASVs}
\usage{
commonASV(x, taxa, n = 50)
}
\arguments{
\item{x}{The community table with Taxon reads counts}

\item{n}{Interger, Number of top ASVs to retrieve}
}
\value{
Two data.frames
\item{by_samp}{data.frame with the top n ASVs by presence in samples }
\item{by_reads}{data.frame with the top n ASVs by number of reads}
}
\description{
Retrieve the most common ASVs by total number of reads and total number of samples present
}
\details{
This function will order the ASVs by total number of reads and by
number of samples present and return two lists containing the of the n top ASVs
}
\examples{

}
\seealso{
\code{\link{ASV_outliers}}
}
\author{
Pedro Martinez Arbizu
}
