% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countTaxa.r
\name{countTaxa}
\alias{countTaxa}
\title{Count ASVs/OTUs per Taxon and Area}
\usage{
countTaxa(x, taxa, fac, ...)
}
\arguments{
\item{x}{The community table with Taxon reads counts}

\item{taxa}{vector with taxon information}

\item{fac}{vector of factors to discriminate}
}
\value{
A data frame
}
\description{

}
\details{

}
\examples{
countTaxa(deepMeio[,9:27],deepMeio$Group,fac= c(rep('area1',7),rep('area2',12)))




}
\seealso{
\code{\link{barp.table}}
}
\author{
Pedro Martinez Arbizu
}
