% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2pp-package.R
\docType{package}
\name{dada2pp-package}
\alias{dada2pp}
\alias{dada2pp-package}
\title{dada2pp: Miscelaneous Functions for Post-Processing dada2 Results.}
\description{
Useful functions to analyse and plot metabarcoding data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pmartinezarbizu/dada2pp}
}

}
\author{
\strong{Maintainer}: Pedro Martinez Arbizu \email{pmartinez@senckenberg.de}

}
\keyword{internal}
