% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.tt.r
\name{divers.tt}
\alias{divers.tt}
\title{Diversity Values Based on Taxon Table}
\usage{
divers.tt(x, by, sample = min(N) * 0.8, pal)
}
\arguments{
\item{x}{}

\item{by}{}

\item{sample}{}

\item{pal}{}
}
\value{
An object of class divers_tt
}
\description{

}
\details{

}
\examples{



}
\seealso{
\code{\link{plot.divers_tt}}
}
\author{
Pedro Martinez Arbizu
}
