% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diverse.tt.r
\name{diverse.tt}
\alias{diverse.tt}
\title{Diversity Values Based on Taxon Table}
\usage{
diverse.tt(x, by, sample = min(N) * 0.8, pal)
}
\arguments{
\item{x}{community table}

\item{by}{aggregating vector}

\item{sample}{number of n for rarefaction (default is min(N)*0.8)}

\item{pal}{Palette object from pal2table}
}
\value{
An object of class divers_tt
This object has a plot method
}
\description{
Calculate miscelaneous diversity values
}
\details{
Returns parameters N= number of specimens, S= number of species,
H= Shannon index, J= Pielou\'s evenness and ES= estimated number od species for N =sample
}
\examples{



}
\seealso{
\code{\link{plot.diverse_tt}}
}
\author{
Pedro Martinez Arbizu
}
