% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiadonis.r
\name{multiadonis}
\alias{multiadonis}
\title{Multiple Adonis Test for Taxon Table}
\usage{
multiadonis(datasets, fac, ...)
}
\arguments{
\item{datasets}{a list of datasets}

\item{fac}{factor for adonis (groups)}
}
\value{
A dataframe with results of adonis
}
\description{

}
\details{

}
\examples{
datasets <- list('meio'=meio[,9:27],'crust'=crust[,9:27],'NeHaCyMi'=NeHaCyMi[,9:27], 'worms'=worms[,9:27])
multiadonis(datasets,area)



}
\seealso{
\code{\link{adonis}}
}
\author{
Pedro Martinez Arbizu
}
