% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot2census.r
\name{pivot2census}
\alias{pivot2census}
\title{From Community Table to Census Format}
\usage{
pivot2census(x, y, dat)
}
\description{
From Community Table to Census Format
}
\author{
Pedro Martinez Arbizu
}
