% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.barp_table.r
\name{plot.barp_table}
\alias{plot.barp_table}
\title{Method plot for object barp_table}
\usage{
\method{plot}{barp_table}(x, main = "dada2sum", pal2table, oma1 = 12,
  ncol.leg = ceiling(nrow(x$table)/5), x.leg = -ncol(x$table) * 1.5,
  y.leg = -0.5, pt.cex = 2.5, x.int = 1)
}
\arguments{
\item{x}{Object of class barp_table}

\item{main}{One of 'vsearchsum' or 'dada2sum' or a vector with 4 elements which are the titles the plots}

\item{pal2table}{Object of class pal2table for matching colors to Taxa}

\item{oma1}{Space on lower margin to allocate legend (passed to graphical parameter oma)}

\item{ncol.leg}{Number of columns of legend}

\item{x.leg}{X value for upper left corner of legend}

\item{y.leg}{Y value for upper left corner of legend}

\item{pt.cex}{Size of symbol in legend}

\item{x.int}{Valeue passd to x.intersp (see ?legend )}
}
\value{
A graph with 4 barplots
}
\description{
Plots 4 barplots in one page using a barp_table object as input
}
\details{
This takes a table created by barp.table and creates 4 barplots
whith absolute and relative number of reads
and absolute and relative number of OTUs/ASV per Taxon
}
\examples{

bp1 <- barp.table(meio[,9:27],meio$Group,sum)
colmeio <- pal2table( c(rownames(bp1$table),'Branchiopoda','Calanoida','Cyclopoida','Harpacticoida','Isopoda','Misophrioida','Ostracoda','Siphonostomatoida','Tanaidacea','Tantulocarida'),pal='c25bro2')
plot(bp1,pal2table=colmeio)

}
\seealso{
\code{\link{plot.barp_table}}
}
\author{
Pedro Martinez Arbizu
}
