% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.divers_tt.r
\name{plot.divers_tt}
\alias{plot.divers_tt}
\title{Plot Method for divers.tt}
\usage{
\method{plot}{divers_tt}(x)
}
\arguments{
\item{x}{}
}
\value{
A graph with 4 boxplots
}
\description{

}
\details{

}
\examples{



}
\seealso{
\code{\link{plot.divers_tt}}
}
\author{
Pedro Martinez Arbizu
}
