% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pal2table.r
\name{plot.pal2table}
\alias{plot.pal2table}
\title{Plots pal2table object}
\usage{
\method{plot}{pal2table}(x)
}
\arguments{
\item{x}{An object created with pal2table}
}
\description{
Plots symbols, colors and the corresponding factor labels of a pal2table object.
}
\examples{
data(iris)
irc <- pal2table(iris$Species)
plot(irc)

}
\seealso{
\code{\link{pal2table}}
}
\author{
Pedro Martinez Arbizu
}
