% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefyASV.r
\name{rarefyASV}
\alias{rarefyASV}
\title{rarefy and Plot ASV reads}
\usage{
rarefyASV(x, step)
}
\arguments{
\item{x}{ASV table}

\item{step}{Step used to rarefy}
}
\value{
A plot
}
\description{
rarefy and Plot ASV reads
}
\examples{

rr <- rarefyASV(deepMeio[,9:ncol(deepMeio)],step=200)
plot(rr)

#provide colors for groups of samples 
plot(rr,col=c(rep('red',10),rep('green',9)))

}
\author{
Pedro Martinez Arbizu
}
