% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadLengthPlot.r
\name{ReadLengthPlot}
\alias{ReadLengthPlot}
\title{Plot ASV read length frequency}
\usage{
ReadLengthPlot(
  x,
  xlab = "read length",
  ylab = "number of  ASVs",
  mode.lab = "bp",
  pos = 2,
  ...
)
}
\arguments{
\item{x}{The column with read length information}
}
\value{
A plot
}
\description{

}
\details{

}
\examples{
countTaxa(deepMeio[,9:27],deepMeio$Group,fac= c(rep('area1',7),rep('area2',12)))




}
\seealso{

}
\author{
Pedro Martinez Arbizu
}
