% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxFromWorms.r
\name{taxFromWorms}
\alias{taxFromWorms}
\title{Retrieves Taxonomy from World Register of Marine Species}
\usage{
taxFromWorms(x, ...)
}
\arguments{
\item{x}{A vector of species names}

\item{...}{additional arguments passed to worms::wormsmatchsp and worms::wormsmatchgen.}
}
\description{
Query a vector to taxa names against WORMS and retrives valid names and classification
}
\details{
The taxonomic assignment of the ASVs resulting from blasting against GenBank is not allways up to date. This fuction will query the names 
against Worl Register of Marine Species at https://www.marinespecies.org/ by using the package worms.
The function returns a data.frame with the columns "kingdom", "phylum", "class", "order",
"family", "genus", "AphiaID", "status", "scientificname", "valid_name" and"query_name".

The function assumes that the separator between genus name and species name is either a space or an underscore.
}
\examples{
data(deepMeio)
tax_deepMeio <- taxFromWorms(deepMeio$Species[1:51])

#example retrieve non only marine species 
tax_deepMeio <- taxFromWorms(deepMeio$Species[1:51], marine_only=FALSE)



}
\seealso{
\code{\link{worms}}
}
\author{
Pedro Martinez Arbizu
}
