% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackki.R
\name{jackki}
\alias{jackki}
\title{Jackknife for Adonis}
\usage{
jackki(x, data, strata = NULL, permutations = 1, ...)
}
\arguments{
\item{x}{Model formula. The LHS is either community matrix or dissimilarity matrix (eg. from vegdist or dist)
See adonis() for details. The RHS are factors that must be column names of a data frame specified with argument data.}

\item{data}{The data frame of independent variables having as column names the factors specified in formula.}

\item{strata}{String. The name of the column with factors to be used as strata.}

\item{permutations}{Numer of permutation. Default is 1, as we are only interested in F value, not p value.}

\item{(...)}{Any other parameter passed to adonis}
}
\value{
An object of class jackki.

\itemize{
\item res: Table showing the results of adonis for each excluded species its final classification.
\item jackknife: unbiased F estimate, unbiased variance of F, upper and lower 95\% confidence intervals.
}

The plot method displays a dotchart with the adonis F value for every excluded species.
A grey solid line represents the unbiased estimate of F and a dotted line to the left and to the rigth of it
the 95\% confidence intervals.
Segregating species use triangle as symbol, n.s. species a square and aggregating species a circle.
Plot can be confined to segregating and aggregating species only by using reduce = TRUE.
Graphical parameters can be passed to dotchart().
}
\description{
Calculates the relative importance of variables using jackknife
}
\details{
Community ecologist want to know how species contribute to dissimilarity between groups.
The routine simper() is commonly used to achive this task.
But simper is limmited to decomposition of the bray-curtis dissimilarity.
jackki() offers a more flexible alternative by analysing the relative contribution of each species 
to observed adonis difference (value F.Model). adonis is calculated
by leaving out one species each time and the resulting F value is recorded.
The unbiased F estimate and 95\% confidence intervals are used
to classify the species into following categories
\itemize{
\item segregating species: removal results in significantly lower F value
\item aggregating species: removal results in significantly higher F value
\item n.s.: removal does not change observed F value significantly 
}
As we are only interested in F value, permutations are set to 1 by default. 
Set permutations to higher value if you want to record the effect of species removal to p value.
}
\examples{

library(vegan)
data(dune)
data(dune.env)
x <- jackki(dune~Management,data=dune.env)

plot(x)

}
\seealso{
\code{\link{simper}} \code{\link{adonis}} \code{\link{dotchart}}
}
\author{
Pedro Martinez Arbizu
}
