% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackki.R
\name{jackki}
\alias{jackki}
\title{Jackknife for Adonis}
\usage{
jackki(x, data, strata = NULL, permutations = 1, ...)
}
\arguments{
\item{x}{Model formula. The LHS is either community matrix or dissimilarity matrix (eg. from vegdist or dist)
See adonis() for details. The RHS are factors that nmust be column names of a data.frame specified with argument data.}

\item{data}{The data frame of independent varibles having as column names the factors specified in formula.}

\item{strata}{String. The name of the column with factors to be used as strata.}

\item{permutations}{Numer of permutation. Default is 1, as we are only interested in F value, not p value.}

\item{(...)}{Any other parameter passed to adonis}
}
\value{
An object of class jackki.

\describe{
\item{res}{results} 
\item{jackknife}{kckko res}
}
}
\description{
Calculates the relative importance of varibles using jackknife
}
\details{
Community ecologist want to know how species contribute to dissimilarity between groups.
The routine simper() is commonly used to achive this task.
But simper is limmited to decomposition of the bray-curtis dissimilarity.
jackki() offers a more flexible alternative by analysing the relative contribution of each species 
to observed adonis() difference (value F.Model). adonis is calculated
and the resulting F value recorded by leaving on species out.
The unbiased F estimate and confidence 95% intervals are used
to classify the species into following categories
\describe{
\item{disgregating species}{removal results in significantly lower F value}
\item{aggegating species}{removal results in in significantly higher F value}
\item{n.s.}{removal does not change observed F value significantly} 
}
As we are only interested in F value, permutations are set to 1 by default. 
Set permutations to higher value if you want to record the effect of species removal to p value.
}
\examples{



}
\seealso{
\code{\link{plot.barp_table}}
}
\author{
Pedro Martinez Arbizu
}
