% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.diverse_tt.r
\name{plot.diverse_tt}
\alias{plot.diverse_tt}
\title{Plot Method for divers.tt}
\usage{
\method{plot}{diverse_tt}(x, ty = "ASVs")
}
\arguments{
\item{x}{}
}
\value{
A graph with 4 boxplots
}
\description{

}
\details{

}
\examples{



}
\seealso{
\code{\link{plot.diverse_tt}}
}
\author{
Pedro Martinez Arbizu
}
