% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefyASV.r
\name{rarefyASV}
\alias{rarefyASV}
\title{rarefy and Plot ASV reads}
\usage{
rarefyASV(x, step)
}
\arguments{
\item{x}{ASV table}

\item{step}{Step used to rarefy}
}
\value{
A plot
}
\description{

}
\details{

}
\examples{




}
\seealso{

}
\author{
Pedro Martinez Arbizu
}
